// src/Domain/Dtos/Filters/Enums/FilterOperator.ts
enum FilterOperator {
  EqualsTo = "EqualsTo",
  NotEqualsTo = "NotEqualsTo",
  StartsWith = "StartsWith",
  NotStartsWith = "NotStartsWith",
  Contains = "Contains",
  NotContains = "NotContains",
  EndsWith = "EndsWith",
  NotEndsWith = "NotEndsWith",
  GreaterThan = "GreaterThan",
  GreaterThanOrEqualTo = "GreaterThanOrEqualTo",
  LessThan = "LessThan",
  LessThanOrEqualTo = "LessThanOrEqualTo",
}

export default FilterOperator;